requirements.txt
daphne
channels
social-auth-app-django
social-auth-core
requests
requests-oauthlib
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
pytz
Pillow
celery
redis
django-celery-beat
twilio